New test for GetVariable() with Attribute valid
This is a new test of GetVariable() when
Attributes is not NULL. DataSize is 0, Data is NULL

In this case the GetVariable() is expected to return EFI_BUFFER_TOO_SMALL,
with Attributes updated with the attribute value set for this variable.

There is a provision in the spec to get only the variable Attributes using the GetVariable() function.
This is a case where the data to store the variable is not passed and DataSize is set to 0

This is newly added in UEFI v2.8
Reference: https://mantis.uefi.org/mantis/view.php?id=1954

This is related to Issue: https://github.com/tianocore/edk2-test/issues/260